requirements.txt
streamlit>=1.32
yt-dlp>=2024.3.10
youtube-transcript-api>=0.6.2
python-docx>=1.1.0
